align DataFrame and Series gmap with underlying data

+++ pandas/io/formats/style.py
@@ -1525,8 +1525,8 @@ class Styler:
         if gmap is None:  # the data is used the gmap
             gmap = s.to_numpy(dtype=float)
         else:  # gmap is conformed to the data shape
-            if isinstance(gmap, (Series, DataFrame)):
-                gmap = gmap.reindex_like(s, method=None).to_numpy()  # align indx / cols
+            if isinstance(gmap, (Series, DataFrame)):  # align indx / cols
+                gmap = gmap.reindex_like(s, method=None).to_numpy(dtype=float)
             else:
                 gmap = np.asarray(gmap, dtype=float)
                 if gmap.shape != s.shape:  # check valid input
